// client/deadletter/app/schemas/DevDataSchema.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type Id = string;
/**
 * Creation date
 */
export type Createdat = string;
/**
 * From Project ID
 */
export type Fromproject = string;
/**
 * List of consumer project IDs
 */
export type Projectconsumers = string[];

export interface DevData {
  _id: Id;
  data: Data;
  createdAt: Createdat;
  fromProject: Fromproject;
  projectConsumers?: Projectconsumers;
}
/**
 * Data field
 */
export interface Data {
  [k: string]: unknown;
}